+++ apps/api/src/app/app.module.ts
@@ -20,4 +20,4 @@ import { AppService } from './app.service'
   controllers: [AppController],
   providers: [AppService],
 })
-export class CodelabAppModule {}
+export class AppModule {}
