+++ libs/backend/src/application/auth/config/auth.tokens.ts
@@ -1,0 +1,3 @@
+export class AuthTokens {
+  public static readonly Auth0Config = 'auth0'
+}
