+++ libs/backend/src/application/auth/index.ts
@@ -1,3 +1,4 @@
 export * from './auth.module'
-export * from './graphql/GqlAuthGuard'
+export * from './config/auth.tokens'
+export * from './graphql/gql-auth.guard'
 export * from './interfaces/jwt.interface'
