+++ libs/backend/src/application/auth/jwt.strategy.ts
@@ -3,7 +3,8 @@ import { ConfigService } from '@nestjs/config'
 import { PassportStrategy } from '@nestjs/passport'
 import { passportJwtSecret } from 'jwks-rsa'
 import { ExtractJwt, Strategy } from 'passport-jwt'
-import { Auth0Configuration } from './config/authConfig'
+import { Auth0Configuration } from './config/auth.config'
+import { AuthTokens } from './config/auth.tokens'
 import { JwtPayload } from './interfaces/jwt.interface'
 
 @Injectable()
@@ -16,13 +17,14 @@ export class JwtStrategy extends PassportStrategy(Strategy) {
         jwksRequestsPerMinute: 5,
         jwksUri: new URL( //Use the URL helper class, because it's better than relying on the issuer url to not have a trailing /
           '/.well-known/jwks.json',
-          config.get<Auth0Configuration>('auth0')?.issuer,
+          config.get<Auth0Configuration>(AuthTokens.Auth0Config)?.issuer,
         ).href,
       }),
 
       jwtFromRequest: ExtractJwt.fromAuthHeaderAsBearerToken(),
-      audience: config.get<Auth0Configuration>('auth0')?.clientId,
-      issuer: config.get<Auth0Configuration>('auth0')?.issuer,
+      audience: config.get<Auth0Configuration>(AuthTokens.Auth0Config)
+        ?.clientId,
+      issuer: config.get<Auth0Configuration>(AuthTokens.Auth0Config)?.issuer,
       algorithms: ['RS256'],
     })
   }
