+++ libs/backend/src/infrastructure/dgraph/config/dgraph.config.ts
@@ -1,12 +1,13 @@
 import { registerAs } from '@nestjs/config'
 import { get } from 'env-var'
+import { DgraphTokens } from './dgraph.tokens'
 
 export interface DgraphConfiguration {
   endpoint: string
 }
 
 export const dgraphConfig = registerAs(
-  'dgraph',
+  DgraphTokens.DgraphConfig,
   (): DgraphConfiguration => ({
     endpoint: get('CODELAB_DGRAPH_ENDPOINT').required(true).asString(),
   }),
