+++ libs/backend/src/infrastructure/dgraph/config/dgraph.tokens.ts
@@ -1,0 +1,3 @@
+export class DgraphTokens {
+  public static readonly DgraphConfig = 'dgraph'
+}
