+++ libs/backend/src/infrastructure/dgraph/dgraph.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
 import { DgraphClient, DgraphClientStub } from 'dgraph-js-http'
-import { DgraphConfiguration } from './config/dgraphConfig'
+import { DgraphConfiguration } from './config/dgraph.config'
+import { DgraphTokens } from './config/dgraph.tokens'
 
 //dgraph-js gave me this issue for some reason https://discuss.dgraph.io/t/error-13-internal-received-rst-stream-with-code-2-internal-server-error/11957/3
 //using dgraph-js-http for now
@@ -18,7 +19,9 @@ export class DGraphService {
   }
 
   constructor(configService: ConfigService) {
-    const config = configService.get<DgraphConfiguration>('dgraph')
+    const config = configService.get<DgraphConfiguration>(
+      DgraphTokens.DgraphConfig,
+    )
 
     if (!config) {
       throw new Error('No dgraph config found')
