--- libs/modules/user-api/src/application/index.ts
@@ -1,1 +1,0 @@
-export * from './user.resolver'
