--- libs/modules/user-api/src/domain/index.ts
@@ -1,1 +1,0 @@
-export * from './models/User'
