+++ libs/modules/user-api/src/index.ts
@@ -1,3 +1,2 @@
-export * from './application'
-export * from './domain'
+export * from './use-cases'
 export * from './user.module'
