+++ libs/modules/user-api/src/use-cases/create-user/create-user.input.ts
@@ -3,9 +3,5 @@ import { Field, InputType } from '@nestjs/graphql'
 @InputType()
 export class CreateUserInput {
   @Field()
-  email: string
-
-  constructor(email: string) {
-    this.email = email
-  }
+  declare email: string
 }
