+++ libs/modules/user-api/src/use-cases/create-user/index.ts
@@ -1,0 +1,2 @@
+export * from './create-user.input'
+export * from './create-user.service'
