+++ libs/modules/user-api/src/use-cases/index.ts
@@ -1,0 +1,1 @@
+export * from './create-user'
