+++ libs/modules/user-api/src/user.model.ts
@@ -1,17 +1,12 @@
-import { Field, ObjectType } from '@nestjs/graphql'
+import { Field, ID, ObjectType } from '@nestjs/graphql'
 
 @ObjectType()
 export class User {
-  @Field()
-  id: string
+  @Field(() => ID)
+  declare id: string
 
   @Field()
-  email: string
-
-  constructor(id: string, email: string) {
-    this.id = id
-    this.email = email
-  }
+  declare email: string
 
   //Need to add those as well when we implement them
   // apps: [App!] @hasInverse(field: "owner")
