+++ libs/modules/user-api/src/user.resolver.ts
@@ -1,9 +1,8 @@
 import { GqlAuthGuard } from '@codelab/backend'
 import { Injectable, Scope, UseGuards } from '@nestjs/common'
 import { Args, Query, Resolver } from '@nestjs/graphql'
-import { User } from '../domain'
-import { CreateUserInput } from '../domain/dto/CreateUserInput'
-import { CreateUserService } from '../domain/use-cases/create-user.service'
+import { CreateUserInput, CreateUserService } from './use-cases'
+import { User } from './user.model'
 
 @Resolver(() => User)
 @Injectable({ scope: Scope.REQUEST })
@@ -11,7 +10,7 @@ export class UserResolver {
   constructor(private create: CreateUserService) {}
 
   @Query(() => User)
-  @UseGuards(new GqlAuthGuard())
+  @UseGuards(GqlAuthGuard)
   createUser(@Args('input') input: CreateUserInput) {
     return this.create.execute(input)
   }
