+++ schema.api.graphql
@@ -3,7 +3,7 @@
 # ------------------------------------------------------
 
 type User {
-  id: String!
+  id: ID!
   email: String!
 }
